# api/swagger/swagger.json
{"swagger":"2.0","info":{"version":"","title":""},"basePath":"/","schemes":["https","http"],"consumes":["application/json"],"produces":["application/json"],"definitions":{},"paths":{"/version":{"get":{"tags":["default"],"summary":"GET - /version","produces":["application/json"],"responses":{"200":{"description":""}}}},"/v1/stores":{"get":{"tags":["stores"],"summary":"GET - /v1/stores","produces":["application/json"],"responses":{"200":{"description":""}},"parameters":[{"name":"id","in":"query","description":" positive","type":"number","required":false},{"name":"name","in":"query","description":"","type":"string","required":false},{"name":"address","in":"query","description":"","type":"string","required":false},{"name":"phone","in":"query","description":"","type":"string","required":false}]},"post":{"tags":["stores"],"summary":"POST - /v1/stores","produces":["application/json"],"responses":{"200":{"description":""}},"parameters":[{"name":"body","in":"body","schema":{"properties":{"id":{"type":"number","description":" positive"},"name":{"type":"string","description":""},"address":{"type":"string","description":""},"phone":{"type":"string","description":""}}}}]}},"/v1/stores/{storeId}/":{"get":{"tags":["stores"],"summary":"GET - /v1/stores/{storeId}/","produces":["application/json"],"responses":{"200":{"description":""}},"parameters":[{"name":"storeId","in":"path","description":"/^[0-9a-fA-F]{24}$/","type":"string","required":true}]},"patch":{"tags":["stores"],"summary":"PATCH - /v1/stores/{storeId}/","produces":["application/json"],"responses":{"200":{"description":""}},"parameters":[{"name":"storeId","in":"path","description":"/^[0-9a-fA-F]{24}$/","type":"string","required":true},{"name":"body","in":"body","schema":{"properties":{"id":{"type":"number","description":" positive"},"name":{"type":"string","description":""},"address":{"type":"string","description":""},"phone":{"type":"string","description":""}}}}]}},"/v1/products":{"get":{"tags":["products"],"summary":"GET - /v1/products","produces":["application/json"],"responses":{"200":{"description":""}},"parameters":[{"name":"name","in":"query","description":"","type":"string","required":false},{"name":"store","in":"query","description":"Main Store, Just Big Store, Small Store, Tiny Store","type":"string","required":false,"enum":["Main Store","Just Big Store","Small Store","Tiny Store"]},{"name":"barcode","in":"query","description":" positive","type":"number","required":false},{"name":"quantity_from","in":"query","description":"","type":"number","required":false},{"name":"quantity_to","in":"query","description":"","type":"number","required":false}]},"post":{"tags":["products"],"summary":"POST - /v1/products","produces":["application/json"],"responses":{"200":{"description":""}},"parameters":[{"name":"body","in":"body","schema":{"properties":{"name":{"type":"string","description":""},"description":{"type":"string","description":""},"quantity":{"type":"number","description":""}}}}]}}}}